[Translation] Move ProviderFactoryTestCase out of the Tests namespace
